pyproject.toml
[tool.poetry]
name = "grpc-minimal"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
delphai-utils = {extras = ["config", "grpc"], version = "^0.2.0"}
mypy = "^0.800"
mypy-protobuf = "^2.4"
delphai-discovery = "^0.1.3"
"ruamel.yaml" = "^0.16.12"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
pep8-naming = "^0.11.1"
flake8 = "^3.8.3"
poethepoet = "^0.8.0"
grpcio-tools = "^1.32.0"
pytest = "^6.2.2"
pytest-sugar = "^0.9.4"
pytest-asyncio = "^0.14.0"
pytest-snapshot = "^0.5.0"
PyYAML = "^5.4.1"


[tool.poe.tasks]
start = "python src/server.py"
codegen = [
  {cmd = "mkdir -p ./src"},
  {cmd = "protodep up"},
  {cmd = "python -m grpc_tools.protoc -I./.vendor -I. --grpc_python_out=./src --mypy_out=./src --python_out=./src ./proto/*.proto"}, 
  {cmd = "python -m grpc_tools.protoc -I./.vendor -I. --grpc_python_out=./src/proto --mypy_out=./src/proto --python_out=./src/proto ./.vendor/**/*.proto"}
]
test = "pytest"

[tool.pytest.ini_options]
addopts = "-p no:warnings"
filterwarnings = [
  "error"
]